Delete record from the db after hiding Snackbar
Implement record deletion from the db after the Snackbar time has expired
using a Timer and a TimerTask

The Timer starts when the Snackbar starts showing.
When the Snackbar time expires, the record is deleted from the database
(in this case, the user has definitely not undo the deletion)
If the user undoes the deletion, the Timer will be stopped and the record won't be deleted.
The record is still removed from the RececlerView by swipe
and added back to the it if the deletion is cancelled (undo)