add the App class extends Application
The App class stores the date that was selected using the DatePickerDialog on the DayActivity.
If we did not select the date using the dialog (for example, we launched the application for the first time or launched it after closing),
then the date field of this class will be null.
If the date field of this class is null, then we will display the current system date on the DayActivity.

add android:name=".App" in the <application> tag in the AndroidManifest.xml