add the App class extends Application
The App class stores the date that was selected using the DatePickerDialog or left-right arrows on the DayActivity
or that was selected using the CalendarActivity.
If we did not select the date using the dialog/arrows on the DayActivity or through the CalendarActivity
(for example, we launched the application for the first time or launched it after closing),
then the date field of this class will be 0 (default value).
If the date field of this class is 0, then we will display the current system date on the DayActivity.

add android:name=".App" in the <application> tag in the AndroidManifest.xml