add DatePickerDialog to the DayActivity
add date displaying on the DayActivity

if the date has not been set earlier through the dialog/arrows on the DayActivity
or through the CalendarActivity, then the current date is displayed, otherwise the previously set date is displayed.

the dialog opens on the date set earlier, otherwise it opens on the current date.

if the selected date in the dialog is not more than the current day,
then the selected date is displayed and saved. Otherwise, the current date is displayed

add tvDate, sdfDayAndMonth to the DayActivity