change records displaying in the DayActivity's onCreate
the DayActivity displays records for the day stored in the App class date field,
that is, for the day selected earlier by the user via the date picker dialog/arrows on the DayActivity or via the CalendarActivity.
If the date has not been selected earlier (App class date field = 0), then the records for the current day are displayed.